finalHotspot/requirements.txt
Django==5.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
redis==5.0.1
django-redis==5.4.0
africastalking==1.2.5
python-decouple==3.8
whitenoise==6.6.0